+++ R/parallel.R
@@ -16,7 +16,7 @@ lightly_parallelize <- function(X, FUN, jobs = 1, ...) {
   if (is.atomic(X)){
     lightly_parallelize_atomic(X = X, FUN = FUN, jobs = jobs, ...)
   } else {
-    mclapply(X = X, FUN = FUN, mc.cores = jobs, ...)
+    safe_mclapply(X = X, FUN = FUN, mc.cores = jobs, ...)
   }
 }
 
@@ -24,8 +24,18 @@ lightly_parallelize_atomic <- function(X, FUN, jobs = 1, ...){
   jobs <- safe_jobs_imports(jobs)
   keys <- unique(X)
   index <- match(X, keys)
-  values <- mclapply(X = keys, FUN = FUN, mc.cores = jobs, ...)
+  values <- safe_mclapply(X = keys, FUN = FUN, mc.cores = jobs, ...)
   values[index]
+}
+
+# Avoid SIGCHLD handler when mc.cores is 1.
+# Could help avoid zeromq interrupted system call errors.
+safe_mclapply <- function(X, FUN, mc.cores, ...){
+  if (mc.cores > 1){
+    parallel::mclapply(X = X, FUN = FUN, mc.cores = mc.cores, ...)
+  } else {
+    lapply(X = X, FUN = FUN, ...)
+  }
 }
 
 # x is parallelism or jobs